pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pyprojet.sdist]
include = ["py.typed"]

[project]
dependencies = [
    "pyserial>=3.5",
    "pyserial-asyncio>=0.6",
]